requirements.txt
altair==5.3.0
annotated-types==0.6.0
attrs==23.2.0
blinker==1.7.0
blis==0.7.11
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
cymem==2.0.8
de-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-3.7.0/de_core_news_sm-3.7.0-py3-none-any.whl#sha256=d88c737eb7eb766f730f6a2dcb99dfcdb81623e1e0d89a9c638a2182ac19c52e
decorator==4.4.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
es-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.7.0/es_core_news_sm-3.7.0-py3-none-any.whl#sha256=61e6e5530941f5880166855f09f60d7e6ba79ec1e8e45f96244bdb1eb169eb1d
filelock==3.13.3
fr-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.7.0/fr_core_news_sm-3.7.0-py3-none-any.whl#sha256=37e9f1f6a278a5138fabdabcc92cc559da917f9b24c76f0adf6758720d7eab10
fsspec==2024.3.1
gensim==4.3.2
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-speech==2.25.1
googleapis-common-protos==1.63.0
googletrans==3.0.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.4.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.22.2
hyperframe==5.2.0
idna==2.10
imageio==2.34.0
imageio-ffmpeg==0.4.9
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
langcodes==3.3.0
langdetect==1.0.9
lxml==5.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
packaging==23.2
pandas==2.2.1
pillow==10.3.0
preshed==3.0.9
proglog==0.1.10
proto-plus==1.23.0
protobuf==4.25.3
pt-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_sm-3.7.0/pt_core_news_sm-3.7.0-py3-none-any.whl#sha256=a3a8eed42c600af17c70ed3c8720ffe2879b5353055e66fe0c998306329052fd
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.17.2
PyMuPDF==1.24.1
PyMuPDFb==1.24.1
python-dateutil==2.9.0.post0
python-docx==1.1.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rfc3986==1.5.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.2
scipy==1.12.0
setuptools==69.2.0
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
streamlit==1.32.2
sympy==1.12
tenacity==8.2.3
thinc==8.2.3
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tornado==6.4
tqdm==4.66.2
transformers==4.39.3
typer==0.9.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wasabi==1.1.2
watchdog==4.0.0
weasel==0.3.4
